requirements.txt

absl-py==1.0.0
addict==2.4.0
aiofiles==0.8.0
altair==4.2.0
altgraph==0.17.2
anyio==3.5.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Command==0.1.0
cycler==0.11.0
Cython==0.29.28
DateTime==4.4
DateTimeRange==1.2.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.5
distlib==0.3.4
dnspython==2.2.0
docutils==0.18.1
EasyProcess==1.1
efficientnet==1.0.0
email-validator==1.1.3
entrypoint2==1.0
entrypoints==0.4
essential-generators==1.0
executing==0.8.3
Faker==13.3.0
filelock==3.6.0
files==1.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Flask==2.0.3
flatbuffers==2.0
fonttools==4.30.0
future==0.18.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-auth==2.6.0
google-pasta==0.2.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.12.0
h5py==3.6.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
imageio==2.16.1
imgaug==0.4.0
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython-genutils==0.2.0
ipython==8.1.1
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
justpy==0.2.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras==2.8.0
keyring==23.5.0
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
matplotlib==3.5.1
mbstrdecoder==1.1.0
mistune==0.8.4
mss==6.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.4
pkginfo==1.8.2
platformdirs==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyclipper==1.3.0.post2
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
pyinstaller-hooks-contrib==2022.1
pyinstaller==4.9
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pyscreenshot==3.0
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
PyWavelets==1.3.0
pywin32-ctypes==0.2.0
pywin32==303
pywinpty==2.0.3
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==32.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
setuptools==58.1.0
Shapely==1.8.1.post1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
sqlalchemy-ag-grid==0.0.3
SQLAlchemy-Utils==0.38.2
SQLAlchemy==1.4.31
stack-data==0.2.0
starlette==0.18.0
streamlit-aggrid==0.2.3.post2
streamlit==1.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow==2.8.0
termcolor==1.1.0
terminado==0.13.2
tesseract==0.1.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.2.9
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
ttkthemes==3.2.2
twine==3.8.0
typepy==1.3.0
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
validators==0.18.2
virtualenv==20.13.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
zope.interface==5.4.0
